Drop any log events coming from the Raven package
When using Logback, if log events are fired whilst within an appender a deadlock can
occur. By dropping any events coming from the Raven package in the SentryAppender, we
can prevent that part of Logback locking up (other appenders would still process those
log events e.g. a FileAppender)

As part of this we also fixed the ThresholdFilter so it was added to the correct place
on the AsyncAppender rather than the SentryAppender